Minor enhancements & cleanup
