--- zkfold-cardano-scripts/src/ZkFold/Cardano/UPLC/AsterizmRelayer.hs
@@ -1,53 +1,0 @@
-{-# LANGUAGE TemplateHaskell #-}
-
-{-# OPTIONS_GHC -Wno-incomplete-uni-patterns #-}
-{-# OPTIONS_GHC -fplugin-opt PlutusTx.Plugin:profile-all #-}
-{-# OPTIONS_GHC -fplugin-opt PlutusTx.Plugin:conservative-optimisation #-}
-{-# OPTIONS_GHC -Wno-unrecognised-pragmas #-}
-
-{-# HLINT ignore "Unused LANGUAGE pragma" #-}
-
-module ZkFold.Cardano.UPLC.AsterizmRelayer where
-
-import           PlutusLedgerApi.V3          as V3
-import           PlutusLedgerApi.V3.Contexts (ownCurrencySymbol, txSignedBy)
-import           PlutusTx                    (CompiledCode, compile, liftCodeDef, unsafeApplyCode)
-import           PlutusTx.AssocMap           (lookup, toList)
-import           PlutusTx.Prelude            (BuiltinUnit, Integer, Maybe (..), Ord (..), check, fmapDefault,
-                                              lengthOfByteString, ($), (&&), (.), (==), (||))
-import           PlutusTx.Trace              (traceError)
-
-
-type RelayerPKH = PubKeyHash
-
--- | Plutus script (minting policy) for posting signed messages on-chain.
-{-# INLINABLE untypedAsterizmRelayer #-}
-untypedAsterizmRelayer :: RelayerPKH -> BuiltinData -> BuiltinUnit
-untypedAsterizmRelayer pkh ctx' = check $ conditionSigned && (conditionBurning || conditionVerifying)
-  where
-    ctx :: ScriptContext
-    ctx = unsafeFromBuiltinData ctx'
-
-    info :: TxInfo
-    info = scriptContextTxInfo ctx
-
-    minted :: Maybe [(TokenName, Integer)]
-    minted = fmapDefault toList . lookup (ownCurrencySymbol ctx) . mintValueToMap $ txInfoMint info
-
-    (tn, amt) = case minted of
-      Just [x] -> x
-      _        -> traceError "Expected exactly one minting action"
-
-    messageHash :: BuiltinByteString
-    messageHash = unsafeFromBuiltinData . getRedeemer $ scriptContextRedeemer ctx
-
-    conditionSigned = txSignedBy info pkh
-
-    conditionBurning = amt < 0
-
-    conditionVerifying = lengthOfByteString messageHash == 32 && tn == TokenName messageHash
-
-asterizmRelayerCompiled :: RelayerPKH -> CompiledCode (BuiltinData -> BuiltinUnit)
-asterizmRelayerCompiled pkh =
-    $$(compile [|| untypedAsterizmRelayer ||])
-    `unsafeApplyCode` liftCodeDef pkh
