+++ zkfold-cardano-scripts/zkfold-cardano-scripts.cabal
@@ -71,8 +71,6 @@ library
   default-language: Haskell2010
   exposed-modules:
     ZkFold.Cardano.UPLC.Asterizm
-    ZkFold.Cardano.UPLC.AsterizmClient
-    ZkFold.Cardano.UPLC.AsterizmRelayer
     ZkFold.Cardano.UPLC.Common
     ZkFold.Cardano.UPLC.ForwardingScripts
     ZkFold.Cardano.UPLC.PlonkupVerifierToken
